pyproject.toml
[tool.poetry]
name = "backup-restore"
version = "0.3.0"
description = "FastAPI backup and restore service for SQL and Neo4j databases"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "api" }, { include = "backend" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
neo4j = "^5.22.0"
# SQL Database support
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
aiosqlite = "^0.19.0"
python-multipart = "^0.0.6"
requests = "^2.32.3"
httpx = "^0.27.0"
cryptography = "^42.0.8"
paramiko = "^3.5.0"
google-api-python-client = "^2.139.0"
google-auth = "^2.33.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"


[tool.pdm]
distribution = false

[project]
name = "backup-restore"
version = "0.3.0"
description = "FastAPI backup and restore service for SQL and Neo4j databases"
authors = [
    {name = "pmichiels", email = "pmichiels@engaige.tech"},
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "pydantic-settings>=2.3.4",
    "python-dotenv>=1.0.1",
    "neo4j>=5.22.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "pydantic[email]>=2.12.4",
    "httpx>=0.27.0",
    "cryptography>=42.0.8",
    "paramiko>=3.5.0",
    "google-api-python-client>=2.139.0",
    "google-auth>=2.33.0",
]
requires-python = ">=3.13,<3.14"
readme = "README.md"
license = {text = "MIT"}

